+++ wsidicom/file/wsidicom_file_source.py
@@ -18,11 +18,10 @@ import io
 import logging
 from collections import defaultdict
 from pathlib import Path
-from typing import BinaryIO, Dict, Iterable, List, Optional, Tuple, Union
+from typing import BinaryIO, Dict, Iterable, List, Optional, Union
 
 from pydicom import dcmread
 from pydicom.errors import InvalidDicomError
-from pydicom.filereader import _read_file_meta_info, read_preamble
 from pydicom.fileset import FileSet
 from pydicom.uid import UID, MediaStorageDirectoryStorage
 
@@ -30,7 +29,8 @@ from wsidicom.errors import (
     WsiDicomNotFoundError,
     WsiDicomNotSupportedError,
 )
-from wsidicom.file.wsidicom_file import WsiDicomFile
+from wsidicom.file.io import WsiDicomReader
+from wsidicom.file.io.wsidicom_io import WsiDicomIO
 from wsidicom.file.wsidicom_file_image_data import WsiDicomFileImageData
 from wsidicom.geometry import Size
 from wsidicom.graphical_annotations import AnnotationInstance
@@ -54,33 +54,33 @@ class WsiDicomFileSource(Source):
             Files to open. Can be a path or stream for a single file, a list of paths or
             streams for multiple files, or a path to a folder containing files.
         """
-        self._level_files: List[WsiDicomFile] = []
-        self._label_files: List[WsiDicomFile] = []
-        self._overview_files: List[WsiDicomFile] = []
-        self._annotation_files: List[BinaryIO] = []
+        self._levels: List[WsiDicomReader] = []
+        self._labels: List[WsiDicomReader] = []
+        self._overviews: List[WsiDicomReader] = []
+        self._annotations: List[WsiDicomIO] = []
         for file in self._list_input_files(files):
             try:
-                stream, filepath = self._open_file(file)
+                stream = self._open_file(file)
                 sop_class_uid = self._get_sop_class_uid(stream)
                 if sop_class_uid == WSI_SOP_CLASS_UID:
                     try:
-                        wsi_file = WsiDicomFile(stream, filepath, filepath is not None)
-                        if wsi_file.image_type == ImageType.VOLUME:
-                            self._level_files.append(wsi_file)
-                        elif wsi_file.image_type == ImageType.LABEL:
-                            self._label_files.append(wsi_file)
-                        elif wsi_file.image_type == ImageType.OVERVIEW:
-                            self._overview_files.append(wsi_file)
+                        reader = WsiDicomReader(stream)
+                        if reader.image_type == ImageType.VOLUME:
+                            self._levels.append(reader)
+                        elif reader.image_type == ImageType.LABEL:
+                            self._labels.append(reader)
+                        elif reader.image_type == ImageType.OVERVIEW:
+                            self._overviews.append(reader)
                     except WsiDicomNotSupportedError:
-                        logging.info(f"Non-supported file {stream.name}.")
-                        if filepath is not None:
+                        logging.info(f"Non-supported file {stream}.")
+                        if stream.filepath is not None:
                             stream.close()
                 elif sop_class_uid == ANN_SOP_CLASS_UID:
-                    self._annotation_files.append(stream)
-                elif filepath is not None:
+                    self._annotations.append(stream)
+                elif stream.filepath is not None:
                     logging.debug(
                         f"Non-supported SOP class {sop_class_uid} "
-                        f"for file {stream.name}."
+                        f"for file {stream}."
                     )
                     # File was opened but not supported SOP class.
                     stream.close()
@@ -88,9 +88,9 @@ class WsiDicomFileSource(Source):
                 logging.error(
                     f"Failed to open file {file.name} due to exception", exc_info=True
                 )
-        if len(self._level_files) == 0:
+        if len(self._levels) == 0:
             raise WsiDicomNotFoundError("Level files", str(files))
-        self._base_dataset = self._get_base_dataset(self._level_files)
+        self._base_dataset = self._get_base_dataset(self._levels)
         self._slide_uids = self._base_dataset.uids.slide
         self._base_tile_size = self._base_dataset.tile_size
 
@@ -103,38 +103,41 @@ class WsiDicomFileSource(Source):
     def level_instances(self) -> Iterable[WsiInstance]:
         """The level instances parsed from the source."""
         return self._create_instances(
-            self._level_files, self._slide_uids, self._base_tile_size
+            self._levels, self._slide_uids, self._base_tile_size
         )
 
     @property
     def label_instances(self) -> Iterable[WsiInstance]:
         """The label instances parsed from the source."""
-        return self._create_instances(self._label_files, self._slide_uids)
+        return self._create_instances(self._labels, self._slide_uids)
 
     @property
     def overview_instances(self) -> Iterable[WsiInstance]:
         """The overview instances parsed from the source."""
-        return self._create_instances(self._overview_files, self._slide_uids)
+        return self._create_instances(self._overviews, self._slide_uids)
 
     @property
     def annotation_instances(self) -> Iterable[AnnotationInstance]:
         """The annotation instances parsed from the source."""
-        return AnnotationInstance.open(self._annotation_files)
+        return [
+            AnnotationInstance.open_dataset(file.read_dataset())
+            for file in self._annotations
+        ]
 
     def close(self) -> None:
-        """Close all opened files in the source. Does not close provided streams."""
-        for image_file in self.image_files:
-            image_file.close()
+        """Close all opened readers in the source. Does not close provided streams."""
+        for reader in self.readers:
+            reader.close()
 
     @property
-    def image_files(self) -> List[WsiDicomFile]:
-        """Return the image files in the source."""
-        file_lists: List[List[WsiDicomFile]] = [
-            self._level_files,
-            self._label_files,
-            self._overview_files,
+    def readers(self) -> List[WsiDicomReader]:
+        """Return the readers in the source."""
+        reader_lists: List[List[WsiDicomReader]] = [
+            self._levels,
+            self._labels,
+            self._overviews,
         ]
-        return [file for file_list in file_lists for file in file_list]
+        return [reader for reader_list in reader_lists for reader in reader_list]
 
     @property
     def is_ready_for_viewing(self) -> Optional[bool]:
@@ -145,7 +148,7 @@ class WsiDicomFileSource(Source):
         """
         if not self.contains_levels:
             return None
-        files = sorted(self.image_files, key=lambda file: file.frame_count)
+        files = sorted(self.readers, key=lambda file: file.frame_count)
         for file in files:
             if file.image_type is None:
                 continue
@@ -160,7 +163,7 @@ class WsiDicomFileSource(Source):
     @property
     def contains_levels(self) -> bool:
         """Returns true source has one level that can be read with WsiDicom."""
-        return len(self.image_files) > 0
+        return len(self._levels) > 0
 
     @classmethod
     def open_dicomdir(cls, path: Union[str, Path]):
@@ -184,11 +187,11 @@ class WsiDicomFileSource(Source):
         return cls(files)
 
     @staticmethod
-    def _open_file(file: Union[Path, BinaryIO]) -> Tuple[BinaryIO, Optional[Path]]:
+    def _open_file(file: Union[Path, BinaryIO]) -> WsiDicomIO:
         """Open stream if file is path. Return stream and optional filepath."""
         if isinstance(file, Path):
-            return open(file, "rb"), file
-        return file, None
+            return WsiDicomIO.open(file, "rb")
+        return WsiDicomIO(file)
 
     @staticmethod
     def _list_input_files(
@@ -229,12 +232,12 @@ class WsiDicomFileSource(Source):
         )
 
     @staticmethod
-    def _get_base_dataset(files: Iterable[WsiDicomFile]) -> WsiDataset:
+    def _get_base_dataset(files: Iterable[WsiDicomReader]) -> WsiDataset:
         """Return file with largest image (width) from list of files.
 
         Parameters
         ----------
-        files: Iterable[WsiDicomFile]
+        files: Iterable[WsiDicomReader]
            List of files.
 
         Returns
@@ -250,14 +253,10 @@ class WsiDicomFileSource(Source):
         )
 
     @staticmethod
-    def _get_sop_class_uid(stream: BinaryIO) -> Optional[UID]:
+    def _get_sop_class_uid(stream: WsiDicomIO) -> Optional[UID]:
         """Return the SOP class UID from file metadata or None if invalid DICOM."""
         try:
-            stream.seek(0)
-            read_preamble(stream, False)
-            metadata = _read_file_meta_info(stream)
-            stream.seek(0)
-            return metadata.MediaStorageSOPClassUID
+            return stream.read_media_storage_sop_class_uid()
         except InvalidDicomError as exception:
             logging.debug(
                 f"Failed to parse DICOM file metadata for file {stream}, not DICOM? "
@@ -268,7 +267,7 @@ class WsiDicomFileSource(Source):
     @classmethod
     def _create_instances(
         cls,
-        files: Iterable[WsiDicomFile],
+        files: Iterable[WsiDicomReader],
         series_uids: SlideUids,
         series_tile_size: Optional[Size] = None,
     ) -> Iterable["WsiInstance"]:
@@ -280,7 +279,7 @@ class WsiDicomFileSource(Source):
 
         Parameters
         ----------
-        files: Sequence[WsiDicomFile]
+        files: Sequence[WsiDicomReader]
             Files to create instances from.
         series_uids: SlideUids
             Uid to match against.
@@ -304,16 +303,16 @@ class WsiDicomFileSource(Source):
 
     @staticmethod
     def _filter_files(
-        files: Iterable[WsiDicomFile],
+        files: Iterable[WsiDicomReader],
         series_uids: SlideUids,
         series_tile_size: Optional[Size] = None,
-    ) -> Iterable[WsiDicomFile]:
+    ) -> Iterable[WsiDicomReader]:
         """
         Filter list of wsi dicom files on uids and tile size if defined.
 
         Parameters
         ----------
-        files: Iterable['WsiDicomFile']
+        files: Iterable[WsiDicomReader]
             Wsi files to filter.
         series_uids: Uids
             Uids to check against.
@@ -322,14 +321,14 @@ class WsiDicomFileSource(Source):
 
         Returns
         ----------
-        List['WsiDicomFile']
+        List[WsiDicomReader]
             List of matching wsi dicom files.
         """
         for file in files:
             if file.dataset.matches_series(series_uids, series_tile_size):
                 yield file
             else:
-                logging.warn(
+                logging.warning(
                     f"{file.filepath} with uids {file.uids.slide} "
                     f"did not match series with {series_uids} "
                     f"and tile size {series_tile_size}"
@@ -338,22 +337,22 @@ class WsiDicomFileSource(Source):
 
     @staticmethod
     def _group_files(
-        files: Iterable["WsiDicomFile"],
-    ) -> Dict[str, List["WsiDicomFile"]]:
+        files: Iterable[WsiDicomReader],
+    ) -> Dict[str, List[WsiDicomReader]]:
         """
         Return files grouped by instance identifier (instances).
 
         Parameters
         ----------
-        files: Iterable[WsiDicomFile]
+        files: Iterable[WsiDicomReader]
             Files to group into instances
 
         Returns
         ----------
-        Dict[str, List[WsiDicomFile]]
+        Dict[str, List[WsiDicomReader]]
             Files grouped by instance, with instance identifier as key.
         """
-        grouped_files: Dict[str, List[WsiDicomFile]] = defaultdict(list)
+        grouped_files: Dict[str, List[WsiDicomReader]] = defaultdict(list)
         for file in files:
             grouped_files[file.uids.identifier].append(file)
         return grouped_files